API/requirements.txt
altgraph==0.17.2
anyio==3.5.0
asgiref==3.5.0
aws==0.2.5
awscrt==0.13.9
awsiotsdk==1.11.0
bcrypt==3.2.0
bidict==0.22.0
boto==2.49.0
boto3==1.22.12
botocore==1.25.12
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.0
docutils==0.18.1
ecdsa==0.17.0
envs==1.4
fabric==2.7.0
fastapi==0.75.2
fastapi-cloudauth==0.4.3
fastapi-socketio==0.0.9
h11==0.13.0
idna==3.3
invoke==1.7.0
Jinja2==3.1.2
jmespath==1.0.0
macholib==1.16
MarkupSafe==2.1.1
netifaces==0.10.6
numpy==1.22.4
paramiko==2.10.4
pathlib2==2.3.7.post1
pip==22.1
prettytable==3.2.0
pyarmor==7.4.2
pyarmor-webui==1.3.3
pyasn1==0.4.8
pycognito==2022.1.0
pycparser==2.21
pydantic==1.9.0
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.3.2
python-jose==3.3.0
python-socketio==5.6.0
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
scipy==1.8.1
setuptools==62.2.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
statistics==1.0.3.5
terminaltables==3.1.10
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
wcwidth==0.2.5
websockets==10.3
wheel==0.37.1
